==> 1.1 data wrangeling.rmd <==
---
title: "1.1 data wrangeling"
author: "TG"
date: "2022-10-30"
output:
   prettydoc::html_pretty:
    theme: cayman
    highlight: github
---


```{r}

library(tidyverse)
library(lubridate)
library(readxl)
library(rfishbase)

```

# load data

```{r}

Aug_22_raw <- read_excel("data/Juvies August 2022.xlsx")
  
June_22_raw <- read_excel("data/Juvies June 2022.xlsx")
   
March_22_raw <- read_excel("data/Juvies March 2022.xlsx")

pre_data_raw<- read_csv("data/transects_data_for_analyses_2018-2020.csv")
```

# start orgenaize

fix date format to be uniform

```{r}
pre_data<-pre_data_raw

pre_data<-rename(.data = pre_data,"line"="...1")

pre_data<-pre_data %>% mutate(new_date = case_when(
  grepl("T00",pre_data$Trip.date) ==T ~ymd_hms(pre_data$Trip.date),
  grepl(" 0:00",pre_data$Trip.date) ==T ~dmy_hm(pre_data$Trip.date)))

pre_data$new_date<-ifelse(is.na(pre_data$new_date),as.character(dmy(pre_data$Trip.date)),
                          as.character(pre_data$new_date))

pre_data$new_date<-ymd(pre_data$new_date)

pre_data<-pre_data %>% relocate(new_date,.after = SiteNo) %>% select(-Trip.date) %>% rename("Trip_date" = "new_date",                                                                     "Site_name"="SiteNo",                                                                         "Sub_site"= "Site",                                                                           "Trans_type"="Letter",
       "Trans_surveyor_id"="ID",
       "Depth" = "Depth.m.",
       "Sp_status" = "c_t")

colnames(pre_data)


pre_data <- pre_data %>% mutate(survey_method = rep("Transects"),.after = month)

```

set Aug to the right format

# Aug 22


```{r}

Aug_22<-Aug_22_raw

```

delete unneccery columns

1. remove all duplicated columns with ".y"
2. calculate objects sizes (A_1,K_1...) and remove the excess columns
3. substrate %
4. column order and names


```{r}


# 1.
duplicated_col<- grep("\\.y$", colnames(Aug_22), ignore.case = T,value = T)

Aug_22<- Aug_22 %>% select(-duplicated_col)

Aug_22[,49:66] <- lapply(Aug_22[,49:66], as.character) #convert all of the A_hight... to numeric
Aug_22[,49:66] <- lapply(Aug_22[,49:66], as.numeric)

Aug_22[,49:66][is.na(Aug_22[49:66])] <- 0

#2.
Aug_22$A1_size <- Aug_22$A1_height *Aug_22$A1_length *Aug_22$A1_width
Aug_22$A2_size <- Aug_22$A2_height *Aug_22$A2_length *Aug_22$A2_width
Aug_22$A3_size <- Aug_22$A3_height *Aug_22$A3_length *Aug_22$A3_width

Aug_22$K1_size = Aug_22$K1_height*Aug_22$K1_width*Aug_22$K1_length
Aug_22$K2_size = Aug_22$K2_height*Aug_22$K2_width*Aug_22$K2_length
Aug_22$K3_size = Aug_22$K3_height*Aug_22$K3_width*Aug_22$K3_length

Aug_22<- Aug_22 %>% select(-c(49:66))

#3.

per_cover_aug = Aug_22[,c(24:48)]

#there are 2 columns for s( S,s), unite:

per_cover_aug[per_cover_aug=="s"] = "S"  

#count how many times a letter appears in the selected columns:
per_cover_aug$S <- apply(per_cover_aug, 1, function(x) length(which(x=="S")))
per_cover_aug$SG <- apply(per_cover_aug, 1, function(x) length(which(x=="SG")))
per_cover_aug$C <- apply(per_cover_aug, 1, function(x) length(which(x=="C")))
per_cover_aug$G <- apply(per_cover_aug, 1, function(x) length(which(x=="G")))
per_cover_aug$B <- apply(per_cover_aug, 1, function(x) length(which(x=="B")))
per_cover_aug$R <- apply(per_cover_aug, 1, function(x) length(which(x=="R")))
per_cover_aug$A <- apply(per_cover_aug, 1, function(x) length(which(x=="A")))
per_cover_aug$SW = NA

per_cover_aug = per_cover_aug %>% mutate(S = (S/25) * 100,
                                 SG = (SG/25) * 100,
                                 C = (C/25) * 100,
                                 G = (G/25) * 100,
                                 B = (B/25) * 100,
                                 R = (R/25) * 100,
                                 A = (A/25) *100)


per_cover_aug = per_cover_aug %>% rowwise() %>% mutate(sum_check = sum(S,SG,C,G,B,R,A))

per_cover_aug = per_cover_aug %>% select(c(S,SG,C,G,B,R,A))

per_cover_aug$sum <- rowSums(per_cover_aug)

per_cover_aug <- per_cover_aug %>% mutate(across(c(1:7), ~ replace(.,sum == 0, NA)))

per_cover_aug$sum<-NULL  


#bind with data
Aug_22 = cbind.data.frame(Aug_22,per_cover_aug)

Aug_22<-Aug_22 %>% select(-c(24:48))


#4.

Aug_22<-Aug_22 %>% select(-Project,-Country,-SiteID,- Expedition)


 Aug_22<- Aug_22 %>% mutate(month = "august",
         year = "2022",
         Trans_No = paste(Transect,Letter))
 
Aug_22<- Aug_22 %>% mutate (observer = ifelse(Observer == "first", 
                           `First Observer.x`, 
                           `Second Observer.x`))

Aug_22<- Aug_22 %>% mutate(observer = case_when(
  Observer == "first" & Letter == "J" ~ `Juveniles First Observer.x`,
  Observer == "first" & Letter != "J" ~ `First Observer.x`,
  Observer == "second" & Letter == "J" ~ `Juveniles Second Observer.x`,
  Observer == "second" & Letter != "J" ~ `Second Observer.x`)) %>% relocate(observer,.after = Date)



Aug_22<- Aug_22 %>% mutate(trans_ID = paste(Location.x,
                                            Site,
                                            Transect,
                                            Date,
                                            `First Observer.x`,
                                            `Second Observer.x`,
                                            `Juveniles First Observer.x`,
                                            `Juveniles Second Observer.x`
                                            ),
                           Trans_surveyor_id = paste(Location.x,Site,Transect,Date,observer))


# remove the observer column (not its name, but whether first/second) and create a column with the number of observers

Aug_22<-Aug_22 %>% select(-c(Observer,
                             `First Observer.x`,
                             `Second Observer.x`,
                             `Juveniles First Observer.x`,
                             `Juveniles Second Observer.x`))  


Aug_22<-Aug_22 %>% 
  group_by(trans_ID,Letter) %>%
  mutate(observer_num_adults = ifelse(n_distinct(observer) > 1 & Letter !="J" ,
                               "two_observers",
                               "one_observer"),
         observer_num_juv = ifelse(n_distinct(observer) > 1 & Letter =="J",
                               "two_observers",
                               "one_observer"),
         Depth = mean(c(as.numeric(`Depth Start`),as.numeric(`Depth End`)),na.rm = T)) %>% # Calculate the mean depth between start and end points of the transect
  ungroup() %>%
  select(-c("Depth End","Depth Start"))


Aug_22<-Aug_22 %>% rename("survey_method" = "Method.x",
                          "Sub_site" = "Site",
                          "Site_name" = "Location.x",
                          "Trans_type" = "Letter",
                          "lat"="Latitude",
                          "lon" = "Longitude")

Aug_22$line<-seq(1:nrow(Aug_22))


Aug_22$meta_to_deployment_id<-NULL
Aug_22$Comment <-rep(NA)

Aug_22<- Aug_22 %>% 
  mutate("Juv_depth_start" = case_when(
  `Juv Transect`=="A" ~ `A Depth Start`,
  `Juv Transect`=="B" ~ `B Depth Start`,
  `Juv Transect`=="C" ~ `C Depth Start`,
  `Juv Transect`=="D" ~ `D Depth Start`,
  `Juv Transect`=="E" ~ `E Depth Start`,
  `Juv Transect`=="F" ~ `F Depth Start`)) %>%
  mutate("Juv_depth_end" = case_when(
  `Juv Transect`=="A" ~ `A Depth End`,
  `Juv Transect`=="B" ~ `B Depth End`,
  `Juv Transect`=="C" ~ `C Depth End`,
  `Juv Transect`=="D" ~ `D Depth End`,
  `Juv Transect`=="E" ~ `E Depth End`,
  `Juv Transect`=="F" ~ `F Depth End`)) %>% 
  mutate("Juv_time_start" = case_when(
  `Juv Transect`=="A" ~ `A Time Start`,
  `Juv Transect`=="B" ~ `B Time Start`,
  `Juv Transect`=="C" ~ `C Time Start`,
  `Juv Transect`=="D" ~ `D Time Start`,
  `Juv Transect`=="E" ~ `E Time Start`,
  `Juv Transect`=="F" ~ `F Time Start`)) %>% 
  mutate("Juv_time_end" = case_when(
  `Juv Transect`=="A" ~ `A Time End`,
  `Juv Transect`=="B" ~ `B Time End`,
  `Juv Transect`=="C" ~ `C Time End`,
  `Juv Transect`=="D" ~ `D Time End`,
  `Juv Transect`=="E" ~ `E Time End`,
  `Juv Transect`=="F" ~ `F Time End`)) %>% select(-(22:45)) 


Aug_22<-Aug_22 %>% group_by(trans_ID,Trans_type) %>% 
  mutate(Depth_juv = mean(c(as.numeric(Juv_depth_start),as.numeric(Juv_depth_end)),na.rm = T)) %>% ungroup() %>% select(-Juv_depth_start,-Juv_depth_end)




```

for now ill remove the species specific columns from pre-data...

```{r}

pre_data<-pre_data %>% select(1:34)

pre_data<-pre_data %>% rename("Date" = "Trip_date",
                              "observer_num_adults"="observer_num",
                              "Amount"="Number_Ind")

pre_data$SW<-NULL
pre_data$Trans_No<-NULL
pre_data$Comment<-NULL


pre_data$sum <- rowSums(pre_data[21:27])


pre_data <- pre_data %>% mutate(across(c(21:27), ~ replace(.,sum == 0, NA)))

pre_data$sum<-NULL


```

```{r}

which(!colnames(pre_data) %in% colnames(Aug_22))

missing_col<-colnames(Aug_22[,(which(!colnames(Aug_22) %in% colnames(pre_data)))])

pre_data <-cbind(pre_data, setNames( lapply(missing_col, function(x) x=NA), missing_col))


which(!colnames(Aug_22) %in% colnames(pre_data))


Aug_22<-Aug_22 %>% select(colnames(pre_data))

Aug_22<- Aug_22 %>%
  relocate(observer_num_juv,.after = observer_num_adults) %>%
  relocate(Object,.after = Length) %>% 
  relocate(`Juv Transect`,.after = Transect) %>%
  relocate(`Time Start`,.after = Depth) %>%
  relocate(`Time End`,.before = Species) %>% 
  relocate(A2_size,.after = A1_size) %>%
  relocate(K3_size,.after = K2_size) %>% 
  relocate(A3_size,.before = K1_size) 

Aug_22<-Aug_22 %>% relocate(46:48,.after = 21)

Aug_22<-Aug_22 %>% relocate(46:47,.after = 36)

Aug_22<-Aug_22 %>% relocate(Notes,.after = Description)

pre_data<-pre_data %>% select(colnames(Aug_22))
```

bind Aug to the previous data

```{r}

Full_data<-rbind(Aug_22,pre_data)
Full_data$Comment<-NULL
Full_data$Trans_No<-NULL
```

ill first bind March and June  and then shape them together to fit the other data

```{r}

June_22<- June_22_raw
March_22<-March_22_raw

colnames(March_22[,(which(!colnames(March_22) %in% colnames(June_22)))])# check which columns not consistent  
colnames(June_22[,(which(!colnames(June_22) %in% colnames(March_22)))]) # check which columns not consistent  

June_22<-June_22 %>%
  select(-Method.y,-Location.y) %>% 
  rename("Location"="Location.x",
         "Method"="Method.x")

March_22$Method = rep("Transects")



March_22 <- March_22 %>% select(colnames(June_22)) # make column in the same order

June_march<- rbind(June_22,March_22)


```

ill repeat more or less the steps of Aug here...

```{r}

duplicated_col<- grep("\\.y$", colnames(June_march), ignore.case = T,value = T)

June_march<- June_march %>% select(-duplicated_col,-(92:109))

June_march[,57:74] <- lapply(June_march[,57:74], as.character)
June_march[,57:74] <- lapply(June_march[,57:74], as.numeric)

June_march[,57:74][is.na(June_march[57:74])] <- 0

#2.
June_march$A1_size <- June_march$A1_height *June_march$A1_length *June_march$A1_width
June_march$A2_size <- June_march$A2_height *June_march$A2_length *June_march$A2_width
June_march$A3_size <- June_march$A3_height *June_march$A3_length *June_march$A3_width

June_march$K1_size = June_march$K1_height*June_march$K1_width*June_march$K1_length
June_march$K2_size = June_march$K2_height*June_march$K2_width*June_march$K2_length
June_march$K3_size = June_march$K3_height*June_march$K3_width*June_march$K3_length


June_march$K1_size[June_march$K1_size==245000]<-0.245 # DATA WAS INPUR IN CM NOT IN M - FIX

June_march<- June_march %>% select(-c(57:74))

#3. calculate cover

per_cover_JM = June_march[,c(32:56)]

#there are 2 columns for s( S,s), unite:

per_cover_JM[per_cover_JM=="s"] = "S"  

#count how many times a letter appears in the selected columns:
per_cover_JM$S <- apply(per_cover_JM, 1, function(x) length(which(x=="S")))
per_cover_JM$SG <- apply(per_cover_JM, 1, function(x) length(which(x=="SG")))
per_cover_JM$C <- apply(per_cover_JM, 1, function(x) length(which(x=="C")))
per_cover_JM$G <- apply(per_cover_JM, 1, function(x) length(which(x=="G")))
per_cover_JM$B <- apply(per_cover_JM, 1, function(x) length(which(x=="B")))
per_cover_JM$R <- apply(per_cover_JM, 1, function(x) length(which(x=="R")))
per_cover_JM$A <- apply(per_cover_JM, 1, function(x) length(which(x=="A")))
per_cover_JM$SW = NA

per_cover_JM = per_cover_JM %>% mutate(S = (S/25) * 100,
                                 SG = (SG/25) * 100,
                                 C = (C/25) * 100,
                                 G = (G/25) * 100,
                                 B = (B/25) * 100,
                                 R = (R/25) * 100,
                                 A = (A/25) *100)


per_cover_JM = per_cover_JM %>% rowwise() %>% mutate(sum_check = sum(S,SG,C,G,B,R,A))

per_cover_JM = per_cover_JM %>% select(c(S,SG,C,G,B,R,A))

per_cover_JM$sum <- rowSums(per_cover_JM)

per_cover_JM <- per_cover_JM %>% mutate(across(c(1:7), ~ replace(.,sum == 0, NA)))

per_cover_JM$sum<-NULL  


#bind with data
June_march = cbind.data.frame(June_march,per_cover_JM)

June_march<-June_march %>% select(-c(32:56))


#4.

June_march<-June_march %>% select(-Project,-Country,-SiteID,- Expedition)

month(June_22_raw$Date[1])

June_march<- June_march %>% mutate(month = case_when(
  month(Date)==6 ~ "june",
  month(Date)==3 ~ "march",
  month(Date)==7 ~ "july"),
         year = "2022",
         Trans_No = paste(Transect,Letter))
 
June_march<- June_march %>% mutate (observer = ifelse(Observer == "first", 
                           `First Observer.x`, 
                           `Second Observer.x`))

June_march<- June_march %>% mutate(observer = case_when(
  Observer == "first" & Letter == "J" ~ `Juveniles First Observer.x`,
  Observer == "first" & Letter != "J" ~ `First Observer.x`,
  Observer == "second" & Letter == "J" ~ `Juveniles Second Observer.x`,
  Observer == "second" & Letter != "J" ~ `Second Observer.x`)) %>% relocate(observer,.after = Date)



June_march<- June_march %>% mutate(trans_ID = paste(Location,
                                            Site,
                                            Transect,
                                            Date,
                                            `First Observer.x`,
                                            `Second Observer.x`,
                                            `Juveniles First Observer.x`,
                                            `Juveniles Second Observer.x`
                                            ),
                           Trans_surveyor_id = paste(Location,Site,Transect,Date,observer))


# remove the observer column (not its name, but whether first/second) and create a column with the number of observers

June_march<-June_march %>% select(-c(Observer,
                             `First Observer.x`,
                             `Second Observer.x`,
                             `Juveniles First Observer.x`,
                             `Juveniles Second Observer.x`))  


June_march<-June_march %>% 
  group_by(trans_ID,Letter) %>%
  mutate(observer_num_adults = ifelse(n_distinct(observer) > 1 & Letter !="J" ,
                               "two_observers",
                               "one_observer"),
         observer_num_juv = ifelse(n_distinct(observer) > 1 & Letter =="J",
                               "two_observers",
                               "one_observer"),
         Depth = mean(c(as.numeric(`Depth Start`),as.numeric(`Depth End`)),na.rm = T)) %>% # Calculate the mean depth between start and end points of the transect
  ungroup() %>%
  select(-c("Depth End","Depth Start"))


June_march<-June_march %>% rename("survey_method" = "Method",
                          "Sub_site" = "Site",
                          "Site_name" = "Location",
                          "Trans_type" = "Letter",
                          "lat"="Latitude",
                          "lon" = "Longitude")

June_march$line<-seq(1:nrow(June_march))


June_march$meta_to_deployment_id<-NULL


June_march<- June_march %>% 
  mutate("Juv_depth_start" = case_when(
  `Juv Transect`=="A" ~ `A Depth Start`,
  `Juv Transect`=="B" ~ `B Depth Start`,
  `Juv Transect`=="C" ~ `C Depth Start`)) %>%
  mutate("Juv_depth_end" = case_when(
  `Juv Transect`=="A" ~ `A Depth End`,
  `Juv Transect`=="B" ~ `B Depth End`,
  `Juv Transect`=="C" ~ `C Depth End`)) %>% 
  mutate("Juv_time_start" = case_when(
  `Juv Transect`=="A" ~ `A Time Start`,
  `Juv Transect`=="B" ~ `B Time Start`,
  `Juv Transect`=="C" ~ `C Time Start`)) %>% 
  mutate("Juv_time_end" = case_when(
  `Juv Transect`=="A" ~ `A Time End`,
  `Juv Transect`=="B" ~ `B Time End`,
  `Juv Transect`=="C" ~ `C Time End`,
  )) %>% select(-(7:18)) 


June_march<-June_march %>% group_by(trans_ID,Trans_type) %>% 
  mutate(Depth_juv = mean(c(as.numeric(Juv_depth_start),as.numeric(Juv_depth_end)),na.rm = T)) %>% ungroup() %>% select(-Juv_depth_start,-Juv_depth_end)


```

```{r}
which(!colnames(Full_data) %in% colnames(June_march))
which(!colnames(June_march) %in% colnames(Full_data))

June_march$Comment<-NULL
June_march$Trans_No<-NULL
June_march$Habitat<-rep(NA)


June_march <- June_march %>% select(colnames(Full_data)) # make column in the same order

Full_data<- rbind(Full_data,June_march)

```

```{r}

Full_data$line<-NULL
Full_data<-Full_data %>% rename('juv_sub_transect'=`Juv Transect`)
Full_data<-Full_data %>% arrange(desc(Date),Site_name)

rm(list=ls()[! ls() %in% c("Full_data","traits")])
```



Upload knolls data:

```{r}
raw_knolls_march <- read_excel("data/Knolls March Juvies 2022.xlsx")
raw_knolls_june <- read_excel("data/Juvies June 2022 knolls.xlsx")
raw_knolls_aug <- read_excel("data/Juvies Knolls August 2022.xlsx")

knolls_march<-raw_knolls_march
knolls_june<-raw_knolls_june
knolls_aug<-raw_knolls_aug

```

For some reason in august 2022 all adults data is NA in transect type, ill change it mannualy here.

```{r}
knolls_aug$Letter[is.na(knolls_aug$Letter)]<-"T"
```


join aug and june

```{r}

which(!colnames(knolls_june) %in% colnames(knolls_aug))
colnames(knolls_aug)[which(!colnames(knolls_aug) %in% colnames(knolls_june))]



missing_col<-colnames(knolls_aug)[which(!colnames(knolls_aug) %in% colnames(knolls_june))]

knolls_june <-cbind(knolls_june, setNames( lapply(missing_col, function(x) x=NA), missing_col))


knolls_june<-knolls_june %>% select(colnames(knolls_aug))

knolls<-bind_rows(knolls_aug,knolls_june)


```

join march:

```{r}
which(!colnames(knolls_march) %in% colnames(knolls))
which(!colnames(knolls) %in% colnames(knolls_march))

knolls_march<-knolls_march %>% select(colnames(knolls))

knolls<-bind_rows(knolls,knolls_march)

```

mutate all knolls data:

```{r}
duplicated_col<- grep("\\.y$", colnames(knolls), ignore.case = T,value = T)

knolls<- knolls %>% select(-duplicated_col)

knolls<- knolls %>% select(-c(45:87)) # remove 0 - 25 and A_hight...k_width


knolls<-knolls %>% select(-Project,-Country,-SiteID,- Expedition)

knolls<- knolls %>% mutate(month = case_when(
  month(Date)==6 ~ "june",
  month(Date)==3 ~ "march",
  month(Date)==7 ~ "july",
  month(Date)==8 ~ "august"),
         year = "2022",
         Trans_No = paste(Transect,Letter))
 
knolls<- knolls %>% mutate (observer = ifelse(Observer == "first", 
                           `First Observer.x`, 
                           `Second Observer.x`))

knolls<- knolls %>% mutate(observer = case_when(
  Observer == "first" & Letter == "J" ~ `Juveniles First Observer.x`,
  Observer == "first" & Letter != "J" ~ `First Observer.x`,
  Observer == "second" & Letter == "J" ~ `Juveniles Second Observer.x`,
  Observer == "second" & Letter != "J" ~ `Second Observer.x`)) %>% relocate(observer,.after = Date)



knolls<- knolls %>% mutate(trans_ID = paste(Location.x,
                                            Site,
                                            Transect,
                                            Date,
                                            `First Observer.x`,
                                            `Second Observer.x`,
                                            `Juveniles First Observer.x`,
                                            `Juveniles Second Observer.x`
                                            ),
                           Trans_surveyor_id = paste(Location.x,Site,Transect,Date,observer))


# remove the observer column (not its name, but whether first/second) and create a column with the number of observers

knolls<-knolls %>% select(-c(Observer,
                             `First Observer.x`,
                             `Second Observer.x`,
                             `Juveniles First Observer.x`,
                             `Juveniles Second Observer.x`))  


knolls<-knolls %>% 
  group_by(trans_ID,Letter) %>%
  mutate(observer_num_adults = ifelse(n_distinct(observer) > 1 & Letter !="J" ,
                               "two_observers",
                               "one_observer"),
         observer_num_juv = ifelse(n_distinct(observer) > 1 & Letter =="J",
                               "two_observers",
                               "one_observer"),
         Depth = mean(c(as.numeric(`Depth Start`),as.numeric(`Depth End`)),na.rm = T)) %>% # Calculate the mean depth between start and end points of the transect
  ungroup() %>%
  select(-c("Depth End","Depth Start"))


knolls<-knolls %>% rename("survey_method" = "Method.x",
                          "Sub_site" = "Site",
                          "Site_name" = "Location.x",
                          "Trans_type" = "Letter",
                          "lat"="Latitude",
                          "lon" = "Longitude")


knolls<-knolls %>% select(-c(13:32))

 

knolls<- knolls %>% 
  mutate("Juv_depth_start" = case_when(
  `Juv Transect`=="A" ~ `A Depth Start`)) %>%
  mutate("Juv_depth_end" = case_when(
  `Juv Transect`=="A" ~ `A Depth End`)) %>% 
  mutate("Juv_time_start" = case_when(
  `Juv Transect`=="A" ~ `A Time Start`)) %>% 
  mutate("Juv_time_end" = case_when(
  `Juv Transect`=="A" ~ `A Time End`)) %>% select(-c(9:12)) # its acually just renaming but i was lazy and copy paste from before


knolls<-knolls %>% group_by(trans_ID,Trans_type) %>% 
  mutate(Depth_juv = mean(c(as.numeric(Juv_depth_start),as.numeric(Juv_depth_end)),na.rm = T)) %>% ungroup() %>% select(-Juv_depth_start,-Juv_depth_end)


knolls$meta_to_deployment_id<-NULL
knolls$Trans_No<- NULL

knolls <- knolls %>% rename("juv_sub_transect"= `Juv Transect`)



which(!colnames(knolls) %in% colnames(Full_data))

which(!colnames(Full_data) %in% colnames(knolls))

missing_col<-colnames(Full_data)[which(!colnames(Full_data) %in% colnames(knolls))]

knolls <-cbind(knolls, setNames( lapply(missing_col, function(x) x=NA), missing_col))

knolls<- knolls %>% select(colnames(Full_data))

```



```{r}
Full_data<- bind_rows(Full_data,knolls)
```


check for missing coordinates,
in cryptic there is no coordinates yet so ill check for J or T


```{r}

coord_check<-Full_data %>% filter(Trans_type != "C",is.na(lat)) %>% distinct(trans_ID,.keep_all = T)

Full_data$lat[Full_data$Site_name=="Igloo"] <-29.503679
Full_data$lon[Full_data$Site_name=="Igloo"] <-34.919032

Full_data$lat[Full_data$trans_ID =="Princess 4 A 2022-07-01 Tal Gavriel Avery Deveto Mai Lazarus Blank"] <-29.494216
Full_data$lon[Full_data$trans_ID =="Princess 4 A 2022-07-01 Tal Gavriel Avery Deveto Mai Lazarus Blank"] <-34.906752

Full_data$lat[Full_data$trans_ID =="IUI 1 A Tal Perevolotsky Sarah Ohayon october 2020"] <-29.500750
Full_data$lon[Full_data$trans_ID =="IUI 1 A Tal Perevolotsky Sarah Ohayon october 2020"] <-34.916592

Full_data$lat[Full_data$trans_ID =="IUI 1 B Tal Perevolotsky Sarah Ohayon october 2020"] <-29.500816
Full_data$lon[Full_data$trans_ID =="IUI 1 B Tal Perevolotsky Sarah Ohayon october 2020"] <-34.916528

Full_data$lat[Full_data$trans_ID =="IUI 1 C Tal Perevolotsky Sarah Ohayon october 2020"] <-29.500816
Full_data$lon[Full_data$trans_ID =="IUI 1 C Tal Perevolotsky Sarah Ohayon october 2020"] <-34.916528

Full_data$lat[Full_data$trans_ID =="Caves 1 B Shevy Rothman Hava Wandel october 2020"] <-29.497676
Full_data$lon[Full_data$trans_ID =="Caves 1 B Shevy Rothman Hava Wandel october 2020"] <-34.914081

Full_data$lat[Full_data$trans_ID =="Dekel 2 B Shevy Rothman Yamit Romano october 2020"] <-29.540998
Full_data$lon[Full_data$trans_ID =="Dekel 2 B Shevy Rothman Yamit Romano october 2020"] <-34.948445

Full_data$lat[Full_data$trans_ID =="Princess 1 A Tal Perevolotsky Yamit Romano october 2020"] <-29.495582
Full_data$lon[Full_data$trans_ID =="Princess 1 A Tal Perevolotsky Yamit Romano october 2020"] <-34.908574

Full_data$lat[Full_data$trans_ID =="Princess 1 B Tal Perevolotsky Yamit Romano october 2020"] <-29.495582
Full_data$lon[Full_data$trans_ID =="Princess 1 B Tal Perevolotsky Yamit Romano october 2020"] <-34.908574

Full_data$lat[Full_data$trans_ID =="japanease_gardens 22 Shira Salingre Shahar Chaikin december 2019"] <- 29.504559 # NEED TO CHECK WITH SHACHAR/SHIRA
Full_data$lon[Full_data$trans_ID =="japanease_gardens 22 Shira Salingre Shahar Chaikin december 2019"] <-34.919521

Full_data$lat[Full_data$trans_ID =="reserve 18 Shira Salingre Shahar Chaikin december 2019"] <-29.510166# NEED TO CHECK WITH SHACHAR/SHIRA
Full_data$lon[Full_data$trans_ID =="reserve 18 Shira Salingre Shahar Chaikin december 2019
"] <-34.924883

Full_data$Depth[Full_data$trans_ID=="Princess 1 A Tal Perevolotsky Yamit Romano october 2020"]<-16


Full_data$lat[Full_data$trans_ID =="Princess 4 A 2022-08-14 Mai Lazarus Blank Mai Lazarus Blank"] <-29.494216
Full_data$lon[Full_data$trans_ID =="Princess 4 A 2022-08-14 Mai Lazarus Blank Mai Lazarus Blank"] <-34.906752



coord_check<-Full_data %>% filter(Trans_type != "C",is.na(lat)) %>% distinct(trans_ID,.keep_all = T)




```

add lat,lon and depth to knolls



```{r}

load("data/knoll_mdata.RData")

knoll_mdata <- knoll_mdata %>% filter(storm == "After") %>%
  select(Site,Knoll,Max_depth,DD_X,DD_Y) %>% 
  mutate(knoll_id = paste(Site,Knoll,sep = "_"))

knoll_mdata<-knoll_mdata %>%
  filter(knoll_id %in% c("Katza_11","Katza_12","Katza_14","Katza_15","Katza_16","Katza_9",
                         "Princess_1","Princess_2"))

colnames(knoll_mdata)<-c("Site_name","Sub_site","Depth.x","lat.x","lon.x","knoll_id" )

knoll_mdata$survey_method<-"Knolls"

knoll_mdata$Sub_site<-as.character(knoll_mdata$Sub_site)
knoll_mdata$lat.x<-as.character(knoll_mdata$lat.x)
knoll_mdata$lon.x<-as.character(knoll_mdata$lon.x)

Full_data<-left_join(Full_data,knoll_mdata)

Full_data<- Full_data %>% mutate(Depth = coalesce(Depth, Depth.x),
                       lat = coalesce(lat,lat.x),
                       lon = coalesce(lon,lon.x))

Full_data<-Full_data %>% select(-c(lat.x,lon.x,Depth.x,knoll_id))

coord_check<-Full_data %>% filter(Trans_type != "C",is.na(lat)) %>% distinct(trans_ID,.keep_all = T)


```



Make the coordinates,depth,and other variables to extent from T to C in the same transect


```{r}


Full_data$adults<- ifelse(Full_data$Trans_type=="J","juv","adu")

Full_data$Trans_type<-as.factor(Full_data$Trans_type)

Full_data$Trans_type<-ordered(Full_data$Trans_type,c("J","T","C"))

Full_data<-Full_data %>% group_by(trans_ID,adults) %>% arrange(desc(Date),Site_name,trans_ID,Trans_type) %>%  fill(lat,lon,Depth,28:34,Description,Photos,Logger,Habitat,Notes,.direction = "down") %>% ungroup()

Full_data<-Full_data %>%
  relocate("Species",.before = "Amount") %>%
  relocate("Confidence",.after = "Length")
```


Unite depth and juv depth to one column


```{r}
Full_data <-Full_data %>% mutate(Depth =coalesce(Depth,Depth_juv),
                               `Time Start` = coalesce(`Time Start`,Juv_time_start),
                               `Time End` = coalesce(`Time End`,Juv_time_end))         

Full_data<- Full_data %>% select(-Juv_time_start,-Juv_time_end,-Depth_juv,-adults) %>% 
  rename("Time_start"=`Time Start`,
         "Time_end" = `Time End`)

Full_data<-Full_data %>% relocate(Habitat,.after = K3_size)


```


set the depth of the igloo and knolls

```{r}

Full_data$Depth[Full_data$Site_name=="Igloo"]<-10.5

Full_data$Depth[Full_data$trans_ID=="Princess 4 A 2022-07-01 Tal Gavriel Avery Deveto Mai Lazarus Blank"]<-4.3

```



start validateing the species names

1. validate and change to fishbase most updated names
2. fix spelling mistakes
3. change species that with mistaken id (due to range for example)
4. recheck validated

```{r}

uniqe_sp<-unique(Full_data$Species)

v_names<-validate_names(uniqe_sp)

uniqe_sp<-as.data.frame(cbind(uniqe_sp,v_names))

```

Fix the spelling mistakes

```{r}


Full_data$Species[Full_data$Species == "Fusigobios neophytus"] <- "Fusigobius neophytus"
Full_data$Species[Full_data$Species == "Dunckerocampus boyelei"] <- "Dunckerocampus boylei"
Full_data$Species[Full_data$Species == "Scorpaenodes paravipinnis"] <- "Scorpaenodes parvipinnis"


uniqe_sp<-unique(Full_data$Species)

v_names<-validate_names(uniqe_sp)

uniqe_sp<-as.data.frame(cbind(uniqe_sp,v_names))

```



compare to the traits data:


```{r}

traits<- read.csv("data\\Belmaker_Species_Traits.csv")

traits<-traits %>% dplyr::filter(Red_Sea == 1)

worng_dis<-v_names[which(!v_names%in%traits$Species_updated_name_fishbase)]

worng_dis<-na.omit(worng_dis)

wn<-traits %>% filter(Species_used_name %in% worng_dis) %>% select(1:3) %>% rename(uniqe_sp=Species_used_name)

```

```{r}

uniqe_sp <- left_join(uniqe_sp,wn)

   
uniqe_sp_2 <- uniqe_sp %>%
  mutate("Species_1" = coalesce(Species_updated_name_fishbase,v_names))

uniqe_sp_2 <- uniqe_sp_2 %>% 
  mutate("Species_2" = coalesce(Species_1,uniqe_sp))
  
  
  
worng_dis[which(!worng_dis %in% wn$uniqe_sp)]



```

Zebrasoma veliferum

fix the last missfits mannualy

```{r}
uniqe_sp_2$Species_2[uniqe_sp_2$uniqe_sp=="Zebrasoma veliferum"]<-"Zebrasoma desjardinii"

uniqe_sp_2<-uniqe_sp_2 %>% select(1,6)

colnames(uniqe_sp_2)<-c("Species","validate_species")

```

Unite and change to the new and validate name

```{r}
Full_data<-left_join(Full_data,uniqe_sp_2)

Full_data <-Full_data %>% select(-Species) %>% relocate(validate_species,.after = Time_end)

Full_data <-Full_data %>% rename(Species=validate_species)





```

remove the observations of "Cantherhines dumerilii" and "Siphamia majimai" that we dont have in the gulf and we dont know which species they could be...



```{r}
Full_data<-Full_data %>% filter(Species != "Cantherhines dumerilii",Species != "Siphamia majimai")


rm(list=ls()[! ls() %in% c("Full_data","traits")])

```



```{r}
#save(Full_data,file ="Full_data.RData")
#write.csv(Full_data,"Full_data.csv")

```

# part 2: traits and size coeff

in this very messy piace of script I added a and b coeff from few different sorces:

1. straight from fish base
2. Bayesian estimate from the transect data of 2018-2021, from file named `ab_eilat` 
3. from the EcoCamp data (need to ass the method or change it later on since i done weird stuff there...)
4. manually fill Bayesian estimate from fishbase to all the missing species

```{r}

#load("Full_data.RData")

ab_eilat<-read_csv("data/ab_eilat.csv")

```



```{r}


Species <- Full_data %>% distinct(Species)


weight <- length_weight(Species$Species,fields = c("Species", "Type","Sex","a","b"))


weight_filter<-weight %>% filter(Type=="TL")


weight_filter<-weight_filter %>% group_by(Species) %>% summarise(a = mean(a),
                                                              b = mean(b),
                                                              convertion_method = rep("mean"))


Species<-left_join(Species,weight_filter)



```


```{r}
ab_eilat$Species<- gsub(".", " ",ab_eilat$Species,fixed=TRUE)

ab_eilat$coeff_a<-as.numeric(ab_eilat$coeff_a)


Species<-left_join(Species,ab_eilat)


Species <- Species %>% mutate(a =coalesce(a,coeff_a),
                              b = coalesce(b,coeff_b),
                              convertion_method = coalesce(convertion_method,note))   

load("data/fish_full_data_ecocamp.RData")

eco_ab <- fish_full_data %>% distinct(Species,.keep_all = T) %>% select(Species,Family,a,b)

colnames(eco_ab)<- c("Species","Family","a_eco","b_eco")

Species<-left_join(Species,eco_ab)

Species <- Species %>% mutate(a =coalesce(a,a_eco),
                              b = coalesce(b,b_eco)) 

Species_no_ab <-  Species %>% filter(is.na(a)) %>% select(1:4)

#write.csv(Species_no_ab,"Species_no_ab.csv")

Species_manual_ab<-read.csv("data/Species_manual_ab.csv")

Species<-left_join(Species,Species_manual_ab)

Species <- Species %>% mutate(a =coalesce(a,a.manu),
                              b = coalesce(b,b.manu),
                              convertion_method = coalesce(convertion_method,convertion_method_manu))


Species<-Species %>% select(1:4)


Full_data<-left_join(Full_data,Species)



```
```{r}

rm(list=ls()[! ls() %in% c("Full_data","traits")])

```

set the column class right


```{r}

str(Full_data)

Full_data$Sub_site<-as.numeric(Full_data$Sub_site)
Full_data$Amount<-as.numeric(Full_data$Amount)
Full_data$Length<-as.numeric(Full_data$Length)
Full_data$year<-as.numeric(Full_data$year)
Full_data$lat<-as.numeric(Full_data$lat)
Full_data$lon<-as.numeric(Full_data$lon)
Full_data$Confidence<-as.numeric(Full_data$Confidence)

```

fix this mistake - all the species are in confidance 1 which is probably a mistake

```{r}
Full_data$Confidence[Full_data$trans_ID=="iui 1 Shira Salingre Shahar Chaikin december 2019"]<-0

```


Understand for which surveys i dont have both transient and cryptic data


```{r}
a<-Full_data %>% distinct(trans_ID,Trans_type)

a<-a %>% group_by(trans_ID) %>% mutate(n=n())


```

# add traits data:

```{r}

#load("Full_data.RData")

```


Ive got few files sources that i need to check and understand if they overlap in species/traits ect.

```{r}
traits_avery<-read.csv("data/traits_for_tal.csv")

my_species<-Full_data %>% distinct(Species)

traits_avery<-traits_avery %>% rename("Species"="species_fishbase_name")

my_species<-left_join(my_species,traits_avery)

trait_matrix<-my_species %>% select(Species,order,family,genus,
                                    home_range,activity,schooling,diet,water_level)

  
  
```


```{r}

Final_traits<-read.csv("data/Final_traits.csv")

Final_traits$Name<-str_to_sentence(Final_traits$Name)
size_traits<-Final_traits %>% select(Name,Size.FishBase,Size_type,Size.Class)
size_traits<-size_traits %>% rename("Species"="Name")

trait_matrix<-left_join(trait_matrix,size_traits)

trait_matrix %>%
  summarise_all(funs(sum(is.na(.))))





```




```{r}
#save(Full_data,file ="Full_data.RData")
#write.csv(Full_data,"Full_data.csv")
#save(trait_matrix,file ="trait_matrix.RData")
#write.csv(trait_matrix,"trait_matrix.csv")

```
